Add retweeting with a cache
To avoid possible rate limits, retweet the existing tweet and place it
in a static dashmap.

Once the mentions timeline is fetched, check the dashmap to see if
there are any already-retweeted tweets.

To prevent a constant loop of tweeting/removing, attach a timestamp
to the tweet id so the cached tweet is removed after a set period
of time.

In addition, add a separate "garbage collection" task which removes all
remaining tweets every 12 hours since some mentioned tweets aren't
retrieved per timeline poll.

The GC is the reason we need a static Dashmap which is concurrent and
shareable across all threads.

Signed-off-by: kingbri <bdashore3@gmail.com>