Refactor systemd service
Due to WSL2 not having systemd support, I did not see that the provided
systemd service will not allow users to input the authorization pin
for the bot to use and causes an infinite restart.

Fix this by hooking the service into a TTY in order to input the appropriate
credentials. The TTY can either be accessed by conspy or chvt.

The README has been updated to reflect the new changes

Signed-off-by: kingbri <bdashore3@gmail.com>